reorganized sidebar nav and other components
