--- src/components/Layouts/Header.js
@@ -1,42 +1,0 @@
-import React from 'react';
-import { makeStyles } from '@material-ui/core/styles';
-import AppBar from '@material-ui/core/AppBar';
-import Toolbar from '@material-ui/core/Toolbar';
-import Typography from '@material-ui/core/Typography';
-import Button from '@material-ui/core/Button';
-import IconButton from '@material-ui/core/IconButton';
-import MenuIcon from '@material-ui/icons/Menu';
-import AccountCircleIcon from '@material-ui/icons/AccountCircle';
-
-const useStyles = makeStyles(theme => ({
-  root: {
-    flexGrow: 1,
-  },
-  title: {
-    flexGrow: 1,
-    color: theme.palette.secondary.main
-  },
-  links: {
-    marginRight: 25
-  },
-}));
-
-export default function Header() {
-  const classes = useStyles();
-
-  return (
-    <div className={classes.root}>
-      <AppBar position="static">
-        <Toolbar>
-          <Typography variant="display1" className={classes.title}>
-            Marist Writing Advisor Reviewer
-          </Typography>
-          <Button color="secondary" className={classes.links}>About</Button>
-          <IconButton edge="start" color="secondary" aria-label="login">
-            <AccountCircleIcon />
-          </IconButton>
-        </Toolbar>
-      </AppBar>
-    </div>
-  );
-}
