+++ src/components/Layouts/SideNav.js
@@ -1,0 +1,151 @@
+
+import React, { Fragment } from 'react';
+import { makeStyles } from '@material-ui/core/styles';
+import CssBaseline from '@material-ui/core/CssBaseline';
+import AppBar from '@material-ui/core/AppBar';
+import { from } from 'rxjs';
+import MenuIcon from '@material-ui/icons/Menu';
+import ChevronLeftIcon from '@material-ui/icons/ChevronLeft';
+import Drawer from '@material-ui/core/Drawer';
+import Toolbar from '@material-ui/core/Toolbar';
+import List from '@material-ui/core/List';
+import Typography from '@material-ui/core/Typography';
+import Divider from '@material-ui/core/Divider';
+import IconButton from '@material-ui/core/IconButton';
+import clsx from 'clsx';
+import { mainListItems, secondaryListItems } from '../Layouts/listItems';
+
+const drawerWidth = 240;
+
+const useStyles = makeStyles(theme => ({
+    root: {
+      display: 'flex',
+    },
+    toolbar: {
+      paddingRight: 24, // keep right padding when drawer closed
+    },
+    toolbarIcon: {
+      background: "#1c1e24",
+      color: "white",
+      display: 'flex',
+      alignItems: 'center',
+      justifyContent: 'flex-end',
+      padding: '0 8px',
+      ...theme.mixins.toolbar,
+    },
+    appBar: {
+      zIndex: theme.zIndex.drawer + 1,
+      transition: theme.transitions.create(['width', 'margin'], {
+        easing: theme.transitions.easing.sharp,
+        duration: theme.transitions.duration.leavingScreen,
+      }),
+    },
+    appBarShift: {
+      marginLeft: drawerWidth,
+      background: "#1c1e24",
+      color: "white",
+      width: `calc(100% - ${drawerWidth}px)`,
+      transition: theme.transitions.create(['width', 'margin'], {
+        easing: theme.transitions.easing.sharp,
+        duration: theme.transitions.duration.enteringScreen,
+  
+      }),
+    },
+    menuButton: {
+      marginRight: 36,
+    },
+    menuButtonHidden: {
+      display: 'none',
+    },
+    title: {
+      flexGrow: 1,
+    },
+    drawerPaper: {
+      background: "#444B58",
+      color: "white",
+      position: 'relative',
+      whiteSpace: 'nowrap',
+      width: drawerWidth,
+      transition: theme.transitions.create('width', {
+        easing: theme.transitions.easing.sharp,
+        duration: theme.transitions.duration.enteringScreen,
+      }),
+    },
+    drawerPaperClose: {
+      overflowX: 'hidden',
+      transition: theme.transitions.create('width', {
+        easing: theme.transitions.easing.sharp,
+        duration: theme.transitions.duration.leavingScreen,
+      }),
+      width: theme.spacing(7),
+      [theme.breakpoints.up('sm')]: {
+        width: theme.spacing(9),
+      },
+    },
+    paper: {
+      padding: theme.spacing(2),
+      display: 'flex',
+      overflow: 'auto',
+      flexDirection: 'column',
+      background: "#444B58",
+      color: "white",
+    },
+    fixedHeight: {
+      height: 240,
+    },
+  }));
+
+const SideNav = () => {
+    const classes = useStyles();
+    const [open, setOpen] = React.useState(true);
+    const handleDrawerOpen = () => {
+      setOpen(true);
+    };
+    const handleDrawerClose = () => {
+      setOpen(false);
+    };
+
+    return (
+        <Fragment>
+            <CssBaseline />
+            
+            <AppBar position="absolute" className={clsx(classes.appBar, !open && classes.appBarShift)}>
+                <Toolbar className={classes.toolbar}>
+                <IconButton
+                    edge="start"
+                    aria-label="open drawer"
+                    onClick={handleDrawerClose}
+                    className={clsx(classes.menuButton, !open && classes.menuButtonHidden)}
+                >
+                    <MenuIcon />
+                </IconButton>
+                <Typography component="h1" variant="h6"  noWrap className={classes.title}>
+                    Writing Advisor Review Tool
+                </Typography>
+                </Toolbar>
+            </AppBar>
+            
+            
+            <Drawer
+                variant="permanent"
+                classes={{
+                paper: clsx(classes.drawerPaper, open && classes.drawerPaperClose),
+                }}
+                open={open}
+            >
+                <div className={classes.toolbarIcon}>
+                <IconButton onClick={handleDrawerOpen}>
+                    <ChevronLeftIcon />
+                </IconButton>
+                </div>
+                <Divider />
+                <List>{mainListItems}</List>
+                <Divider />
+                <List>{secondaryListItems}</List>
+            </Drawer>
+        </Fragment>
+    )
+
+}
+
+export default SideNav;
